Move link-related files out of the transport
