+++ include/zenoh-pico/link/private/manager.h
@@ -19,11 +19,8 @@
 #define TCP_SCHEMA "tcp"
 #define UDP_SCHEMA "udp"
 
-#include "zenoh-pico/utils/private/result.h"
-#include "zenoh-pico/transport/private/link.h"
-
-/*------------------ Result declarations ------------------*/
-_ZN_P_RESULT_DECLARE(_zn_link_t, link)
+#include "zenoh-pico/link/result.h"
+#include "zenoh-pico/link/types.h"
 
 _zn_link_p_result_t _zn_open_link(const char* locator, const clock_t tout);
 void _zn_close_link(_zn_link_t *link);
