+++ include/zenoh-pico/link/result.h
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2017, 2021 ADLINK Technology Inc.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+ * which is available at https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+ *
+ * Contributors:
+ *   ADLINK zenoh team, <zenoh@adlink-labs.tech>
+ */
+
+#ifndef _ZENOH_PICO_LINK_RESULT_H
+#define _ZENOH_PICO_LINK_RESULT_H
+
+#include "zenoh-pico/utils/private/result.h"
+#include "zenoh-pico/link/types.h"
+
+_ZN_P_RESULT_DECLARE(_zn_link_t, link)
+
+#endif /* _ZENOH_PICO_LINK_RESULT_H */
