+++ src/zenoh/link/tcp.c
@@ -12,10 +12,8 @@
  *   ADLINK zenoh team, <zenoh@adlink-labs.tech>
  */
 
-#include <stdlib.h>
 #include "zenoh-pico/system/common.h"
-#include "zenoh-pico/transport/private/manager.h"
-#include "zenoh-pico/utils/private/logging.h"
+#include "zenoh-pico/link/private/manager.h"
 
 _zn_socket_result_t _zn_f_link_tcp_open(void *arg, const clock_t tout)
 {
