+++ src/zenoh/session/utils.c
@@ -25,6 +25,7 @@
 #include "zenoh-pico/session/private/queryable.h"
 #include "zenoh-pico/session/private/query.h"
 #include "zenoh-pico/transport/private/utils.h"
+#include "zenoh-pico/link/private/manager.h"
 #include "zenoh-pico/utils/types.h"
 
 /*------------------ Clone helpers ------------------*/
